refactor(TopicFilterBank): no longer an Island
Its child component can simply be LinkButtons with an `href`

Rename FilterButton to FilterLink
